CAD-1347 cardano-tx-generator:  update to latest node
