+++ benchmarks/shelley3pools/run-tx-generator.sh
@@ -14,7 +14,6 @@ localsock=$BASEDIR/logs/sockets/1
 
 args_common=(
   --config                 ${CONFIGDIR}/configuration-generator.yaml
-  --delegation-certificate ${CONFIGDIR}/genesis/delegation-cert.000.json
   --genesis-file           ${CONFIGDIR}/genesis/genesis.json
   --socket-path            $localsock
   --num-of-txs             $numtx
@@ -27,7 +26,6 @@ args_common=(
   ${TARGETNODES}
   )
 args_byron=(
-  --signing-key            ${CONFIGDIR}/genesis/delegate-keys.000.key
   --sig-key                ${CONFIGDIR}/genesis/delegate-keys.000.key
   --sig-key                ${CONFIGDIR}/genesis/delegate-keys.001.key
   --sig-key                ${CONFIGDIR}/genesis/delegate-keys.002.key
@@ -36,10 +34,9 @@ args_byron=(
   --sig-key                ${CONFIGDIR}/genesis/delegate-keys.005.key
 )
 args_shelley=(
-  --signing-key            ${CONFIGDIR}/genesis/utxo-keys/utxo3.skey
-  --sig-key                ${CONFIGDIR}/genesis/utxo-keys/utxo3.skey
   --sig-key                ${CONFIGDIR}/genesis/utxo-keys/utxo1.skey
-  --sig-key                ${CONFIGDIR}/genesis/utxo-keys/utxo2.skey
+  --sig-key                ${CONFIGDIR}/genesis/utxo-keys/utxo1.skey
+  --sig-key                ${CONFIGDIR}/genesis/utxo-keys/utxo1.skey
 )
 
 echo "starting submission to:  $TARGETNODES"
