+++ cardano-tx-generator/src/Cardano/Benchmarking/GeneratorTx/Submission.hs
@@ -24,24 +24,7 @@
 {-# OPTIONS_GHC -fno-warn-orphans #-}
 
 module Cardano.Benchmarking.GeneratorTx.Submission
-  ( TraceBenchTxSubmit(..)
-  , TraceLowLevelSubmit (..)
-  , ShelleyBlock
-  , BenchTraceConstraints
-  , Params(..)
-  , mkParamsByron
-  , mkParamsShelley
-  , paramsConfig
-  , trBase
-  , trTxSubmit
-  , trConnect
-  , trSubmitMux
-  , trLowLevel
-  , trN2N
-  , submitTx'
-  , NodeToNodeSubmissionTrace(..)
-  , TPSRate(..)
-  , SubmissionParams(..)
+  ( SubmissionParams(..)
   , Submission
   , SubmissionThreadReport
   , mkSubmission
@@ -65,21 +48,17 @@ import qualified Control.Monad.Class.MonadSTM
 import           Control.Monad.Class.MonadTimer (MonadTimer)
 import           Control.Monad.Class.MonadThrow (MonadThrow)
 
-import           Data.Aeson (ToJSON (..), (.=))
-import qualified Data.Aeson as A
 import           Data.ByteString.Lazy (ByteString)
 import qualified Data.List as L
 import qualified Data.List.Extra as L
 import qualified Data.List.NonEmpty as NE
-import           Data.String (String)
 import qualified Data.Text as T
 import           Data.Time.Clock
-                   ( DiffTime, NominalDiffTime, UTCTime)
+                   ( NominalDiffTime, UTCTime)
 import qualified Data.Time.Clock as Clock
 import           Data.Void (Void)
 
 import           Cardano.BM.Tracing
-import           Cardano.BM.Data.Tracer (emptyObject, mkObject, trStructured)
 import           Control.Tracer (Tracer, traceWith)
 
 import qualified Codec.CBOR.Term as CBOR
@@ -93,22 +72,13 @@ import           Cardano.TracingOrphanInstances.Mock()
 import           Cardano.TracingOrphanInstances.Network()
 import           Cardano.TracingOrphanInstances.Shelley()
 
-import           Ouroboros.Consensus.Config.SupportsNode (getNetworkMagic)
-import           Ouroboros.Consensus.Block
-import           Ouroboros.Consensus.Byron.Ledger (ByronBlock (..))
 import           Ouroboros.Consensus.Byron.Ledger.Mempool as Mempool (GenTx)
-import           Ouroboros.Consensus.Shelley.Protocol.Crypto (TPraosStandardCrypto)
-import qualified Ouroboros.Consensus.Shelley.Ledger.Block as Shelley
-import           Ouroboros.Consensus.Config (TopLevelConfig(..))
 import           Ouroboros.Consensus.Ledger.SupportsMempool as Mempool
                    ( GenTxId, HasTxId, TxId, txId, txInBlockSize)
 import           Ouroboros.Consensus.Network.NodeToClient
 import           Ouroboros.Consensus.Node.NetworkProtocolVersion
                   (HasNetworkProtocolVersion (..), nodeToClientProtocolVersion, supportedNodeToClientVersions)
-import           Ouroboros.Consensus.Node.ProtocolInfo (ProtocolInfo (..))
 import           Ouroboros.Consensus.Node.Run (RunNode(..))
-import qualified Ouroboros.Consensus.Cardano as Consensus
-import           Ouroboros.Consensus.Util.Condense (Condense(..))
 
 import           Ouroboros.Network.Mux
                    ( MuxMode(..), OuroborosApplication(..),
@@ -124,492 +94,16 @@ import           Ouroboros.Network.Protocol.TxSubmission.Client (ClientStIdle(..
 import           Ouroboros.Network.Protocol.TxSubmission.Type (BlockingReplyList(..),
                                                                TokBlockingStyle(..),
                                                                TxSizeInBytes)
-import           Ouroboros.Network.NodeToClient (IOManager,
-                                                 NetworkConnectTracers(..),
+import           Ouroboros.Network.NodeToClient (NetworkConnectTracers(..),
                                                  NodeToClientVersionData(..),
                                                  foldMapVersions,
                                                  versionedNodeToClientProtocols)
 import qualified Ouroboros.Network.NodeToClient as NodeToClient
 
-import           Cardano.Config.Logging (LoggingLayer (..))
 import           Cardano.Config.Types (SocketPath(..))
 
-import           Cardano.Benchmarking.GeneratorTx.NodeToNode
-                    (SendRecvConnect,
-                     SendRecvTxSubmission)
+import           Cardano.Benchmarking.GeneratorTx.Params
 
-
-{-------------------------------------------------------------------------------
-  First, cometh the basic types
--------------------------------------------------------------------------------}
-
--- | Transactions not yet even announced.
-newtype UnReqd  tx = UnReqd  [tx]
-
--- | Transactions we decided to announce now.
-newtype ToAnnce tx = ToAnnce [tx]
-
--- | Transactions announced, yet unacked by peer.
-newtype UnAcked tx = UnAcked [tx]
-
--- | Transactions acked by peer.
-newtype Acked tx = Acked [tx]
-
--- | Peer acknowledged this many txids of the outstanding window.
-newtype Ack = Ack Int deriving (Enum, Eq, Integral, Num, Ord, Real)
-
--- | Peer requested this many txids to add to the outstanding window.
-newtype Req = Req Int deriving (Enum, Eq, Integral, Num, Ord, Real)
-
--- | This many Txs sent to peer.
-newtype Sent = Sent Int deriving (Enum, Eq, Generic, Integral, Num, Ord, Real, Show)
-
--- | This many Txs requested by the peer, but not available for sending.
-newtype Unav = Unav Int deriving (Enum, Eq, Generic, Integral, Num, Ord, Real, Show)
-
-newtype TPSRate = TPSRate Double deriving (Eq, Generic, Ord, Show)
-
-instance ToJSON Sent
-instance ToJSON Unav
-instance ToJSON TPSRate
-
-{-------------------------------------------------------------------------------
-  Overall benchmarking trace
--------------------------------------------------------------------------------}
-data TraceBenchTxSubmit txid
-  = TraceBenchTxSubRecv [txid]
-  -- ^ Received from generator.
-  | TraceBenchTxSubStart [txid]
-  -- ^ The @txid@ has been submitted to `TxSubmission`
-  --   protocol peer.
-  | TraceBenchTxSubServAnn [txid]
-  -- ^ Announcing txids in response for server's request.
-  | TraceBenchTxSubServReq [txid]
-  -- ^ Request for @tx@ recieved from `TxSubmission` protocol
-  --   peer.
-  | TraceBenchTxSubServAck [txid]
-  -- ^ An ack (window moved over) received for these transactions.
-  | TraceBenchTxSubServDrop [txid]
-  -- ^ Transactions the server implicitly dropped.
-  | TraceBenchTxSubServOuts [txid]
-  -- ^ Transactions outstanding.
-  | TraceBenchTxSubServUnav [txid]
-  -- ^ Transactions requested, but unavailable in the outstanding set.
-  | TraceBenchTxSubServFed [txid]
-  -- ^ Transactions fed by the feeder.
-  | TraceBenchTxSubServCons [txid]
-  -- ^ Transactions consumed by a submitter.
-  | TraceBenchTxSubIdle
-  -- ^ Remote peer requested new transasctions but none were
-  --   available, generator not keeping up?
-  | TraceBenchTxSubRateLimit DiffTime
-  -- ^ Rate limiter bit, this much delay inserted to keep within
-  --   configured rate.
-  | TraceBenchTxSubSummary SubmissionSummary
-  -- ^ SubmissionSummary.
-  | TraceBenchTxSubDebug String
-  | TraceBenchTxSubError Text
-  deriving (Show)
-
-instance HasSeverityAnnotation (TraceBenchTxSubmit (Mempool.GenTxId blk))
-
-instance HasPrivacyAnnotation (TraceBenchTxSubmit (Mempool.GenTxId blk))
-
-instance Show (GenTxId blk) => Transformable Text IO (TraceBenchTxSubmit (Mempool.GenTxId blk)) where
-  -- transform to JSON Object
-  trTransformer = trStructured
-
-instance {-# OVERLAPS #-} Show (GenTxId blk)
- => ToJSON (Mempool.GenTxId blk) where
-  toJSON txid = A.String (T.pack $ show txid)
-
-instance Show (GenTxId blk)
- => ToObject (TraceBenchTxSubmit (Mempool.GenTxId blk)) where
-  toObject MinimalVerbosity _ = emptyObject -- do not log
-  toObject NormalVerbosity t =
-    case t of
-      TraceBenchTxSubRecv _      -> mkObject ["kind" .= A.String "TraceBenchTxSubRecv"]
-      TraceBenchTxSubStart _     -> mkObject ["kind" .= A.String "TraceBenchTxSubStart"]
-      TraceBenchTxSubServAnn _   -> mkObject ["kind" .= A.String "TraceBenchTxSubServAnn"]
-      TraceBenchTxSubServReq _   -> mkObject ["kind" .= A.String "TraceBenchTxSubServReq"]
-      TraceBenchTxSubServAck _   -> mkObject ["kind" .= A.String "TraceBenchTxSubServAck"]
-      TraceBenchTxSubServDrop _  -> mkObject ["kind" .= A.String "TraceBenchTxSubServDrop"]
-      TraceBenchTxSubServOuts _  -> mkObject ["kind" .= A.String "TraceBenchTxSubServOuts"]
-      TraceBenchTxSubServUnav _  -> mkObject ["kind" .= A.String "TraceBenchTxSubServUnav"]
-      TraceBenchTxSubServFed _   -> mkObject ["kind" .= A.String "TraceBenchTxSubServFed"]
-      TraceBenchTxSubServCons _  -> mkObject ["kind" .= A.String "TraceBenchTxSubServCons"]
-      TraceBenchTxSubIdle        -> mkObject ["kind" .= A.String "TraceBenchTxSubIdle"]
-      TraceBenchTxSubRateLimit _ -> mkObject ["kind" .= A.String "TraceBenchTxSubRateLimit"]
-      TraceBenchTxSubSummary _   -> mkObject ["kind" .= A.String "TraceBenchTxSubSummary"]
-      TraceBenchTxSubDebug _     -> mkObject ["kind" .= A.String "TraceBenchTxSubDebug"]
-      TraceBenchTxSubError _     -> mkObject ["kind" .= A.String "TraceBenchTxSubError"]
-  toObject MaximalVerbosity t =
-    case t of
-      TraceBenchTxSubRecv txIds ->
-        mkObject [ "kind"  .= A.String "TraceBenchTxSubRecv"
-                 , "txIds" .= toJSON txIds
-                 ]
-      TraceBenchTxSubStart txIds ->
-        mkObject [ "kind"  .= A.String "TraceBenchTxSubStart"
-                 , "txIds" .= toJSON txIds
-                 ]
-      TraceBenchTxSubServAnn txIds ->
-        mkObject [ "kind"  .= A.String "TraceBenchTxSubServAnn"
-                 , "txIds" .= toJSON txIds
-                 ]
-      TraceBenchTxSubServReq txIds ->
-        mkObject [ "kind"  .= A.String "TraceBenchTxSubServReq"
-                 , "txIds" .= toJSON txIds
-                 ]
-      TraceBenchTxSubServAck txIds ->
-        mkObject [ "kind"  .= A.String "TraceBenchTxSubServAck"
-                 , "txIds" .= toJSON txIds
-                 ]
-      TraceBenchTxSubServDrop txIds ->
-        mkObject [ "kind"  .= A.String "TraceBenchTxSubServDrop"
-                 , "txIds" .= toJSON txIds
-                 ]
-      TraceBenchTxSubServOuts txIds ->
-        mkObject [ "kind"  .= A.String "TraceBenchTxSubServOuts"
-                 , "txIds" .= toJSON txIds
-                 ]
-      TraceBenchTxSubServUnav txIds ->
-        mkObject [ "kind"  .= A.String "TraceBenchTxSubServUnav"
-                 , "txIds" .= toJSON txIds
-                 ]
-      TraceBenchTxSubServFed txIds ->
-        mkObject [ "kind"  .= A.String "TraceBenchTxSubServFed"
-                 , "txIds" .= toJSON txIds
-                 ]
-      TraceBenchTxSubServCons txIds ->
-        mkObject [ "kind"  .= A.String "TraceBenchTxSubServCons"
-                 , "txIds" .= toJSON txIds
-                 ]
-      TraceBenchTxSubIdle ->
-        mkObject [ "kind" .= A.String "TraceBenchTxSubIdle"
-                 ]
-      TraceBenchTxSubRateLimit limit ->
-        mkObject [ "kind"  .= A.String "TraceBenchTxSubRateLimit"
-                 , "limit" .= toJSON limit
-                 ]
-      TraceBenchTxSubSummary summary ->
-        mkObject [ "kind"    .= A.String "TraceBenchTxSubSummary"
-                 , "summary" .= toJSON summary
-                 ]
-      TraceBenchTxSubDebug s ->
-        mkObject [ "kind" .= A.String "TraceBenchTxSubDebug"
-                 , "msg"  .= A.String (T.pack s)
-                 ]
-      TraceBenchTxSubError s ->
-        mkObject [ "kind" .= A.String "TraceBenchTxSubError"
-                 , "msg"  .= A.String s
-                 ]
-
-instance Show (GenTxId blk) => ToObject (Mempool.GenTxId blk) where
-  toObject MinimalVerbosity _    = emptyObject -- do not log
-  toObject NormalVerbosity _     = mkObject [ "kind" .= A.String "GenTxId"]
-  toObject MaximalVerbosity txid = mkObject [ "kind" .= A.String "GenTxId"
-                                            , "txId" .= toJSON txid
-                                            ]
-
-instance HasSeverityAnnotation (Mempool.GenTxId blk)
-
-instance HasPrivacyAnnotation (Mempool.GenTxId blk)
-
-instance Show (GenTxId blk) => Transformable Text IO (Mempool.GenTxId blk) where
-  trTransformer = trStructured
-
-{-------------------------------------------------------------------------------
-  N2N submission trace
--------------------------------------------------------------------------------}
-instance HasSeverityAnnotation NodeToNodeSubmissionTrace
-instance HasPrivacyAnnotation  NodeToNodeSubmissionTrace
-instance Transformable Text IO NodeToNodeSubmissionTrace where
-  trTransformer = trStructured
-
--- | Low-tevel tracer
-data TraceLowLevelSubmit
-  = TraceLowLevelSubmitting
-  -- ^ Submitting transaction.
-  | TraceLowLevelAccepted
-  -- ^ The transaction has been accepted.
-  | TraceLowLevelRejected String
-  -- ^ The transaction has been rejected, with corresponding error message.
-  deriving (Show)
-
-instance ToObject TraceLowLevelSubmit where
-  toObject MinimalVerbosity _ = emptyObject -- do not log
-  toObject NormalVerbosity t =
-    case t of
-      TraceLowLevelSubmitting -> mkObject ["kind" .= A.String "TraceLowLevelSubmitting"]
-      TraceLowLevelAccepted   -> mkObject ["kind" .= A.String "TraceLowLevelAccepted"]
-      TraceLowLevelRejected m -> mkObject [ "kind" .= A.String "TraceLowLevelRejected"
-                                          , "message" .= A.String (T.pack m)
-                                          ]
-  toObject MaximalVerbosity t =
-    case t of
-      TraceLowLevelSubmitting ->
-        mkObject [ "kind" .= A.String "TraceLowLevelSubmitting"
-                 ]
-      TraceLowLevelAccepted ->
-        mkObject [ "kind" .= A.String "TraceLowLevelAccepted"
-                 ]
-      TraceLowLevelRejected errMsg ->
-        mkObject [ "kind"   .= A.String "TraceLowLevelRejected"
-                 , "errMsg" .= A.String (T.pack errMsg)
-                 ]
-
-instance HasSeverityAnnotation TraceLowLevelSubmit
-
-instance HasPrivacyAnnotation TraceLowLevelSubmit
-
-instance (MonadIO m) => Transformable Text m TraceLowLevelSubmit where
-  -- transform to JSON Object
-  trTransformer = trStructured
-
-type ShelleyBlock = Shelley.ShelleyBlock TPraosStandardCrypto
-
-data BenchTracers m blk =
-  BenchTracers
-  { btBase_       :: Trace  m Text
-  , btTxSubmit_   :: Tracer m (TraceBenchTxSubmit (GenTxId blk))
-  , btConnect_    :: Tracer m SendRecvConnect
-  , btSubmission_ :: Tracer m (SendRecvTxSubmission blk)
-  , btLowLevel_   :: Tracer m TraceLowLevelSubmit
-  , btN2N_        :: Tracer m NodeToNodeSubmissionTrace
-  }
-
--- TODO: move out to Types.hs
-data Params blk where
-  ParamsByron
-    :: TopLevelConfig ByronBlock
-    -> BenchTracers IO ByronBlock
-    -> Params ByronBlock
-  ParamsShelley
-    :: TopLevelConfig ShelleyBlock
-    -> BenchTracers IO ShelleyBlock
-    -> Params ShelleyBlock
-
-tracers :: Params blk -> BenchTracers IO blk
-tracers (ParamsByron   _ trs) = trs
-tracers (ParamsShelley _ trs) = trs
-
-trBase       :: Params blk -> Trace IO Text
-trTxSubmit   :: Params blk -> Tracer IO (TraceBenchTxSubmit (GenTxId blk))
-trConnect    :: Params blk -> Tracer IO SendRecvConnect
-trSubmitMux  :: Params blk -> Tracer IO (SendRecvTxSubmission blk)
-trLowLevel   :: Params blk -> Tracer IO TraceLowLevelSubmit
-trN2N        :: Params blk -> Tracer IO NodeToNodeSubmissionTrace
-trBase       = btBase_       . tracers
-trTxSubmit   = btTxSubmit_   . tracers
-trConnect    = btConnect_    . tracers
-trSubmitMux  = btSubmission_ . tracers
-trLowLevel   = btLowLevel_   . tracers
-trN2N        = btN2N_        . tracers
-
-mkParamsByron
-  :: Consensus.Protocol IO ByronBlock Consensus.ProtocolRealPBFT -> LoggingLayer -> Params ByronBlock
-mkParamsByron ptcl = ParamsByron pInfoConfig . createTracers
- where
-   ProtocolInfo{pInfoConfig} = Consensus.protocolInfo ptcl
-
-mkParamsShelley
-  :: Consensus.Protocol IO ShelleyBlock Consensus.ProtocolRealTPraos -> LoggingLayer -> Params ShelleyBlock
-mkParamsShelley ptcl = ParamsShelley pInfoConfig . createTracers
- where
-   ProtocolInfo{pInfoConfig} = Consensus.protocolInfo ptcl
-
-type BenchTraceConstraints blk =
-  ( Condense (GenTxId blk)
-  , Show (GenTxId blk)
-  , ToJSON (TxId (GenTx blk))
-  , Transformable Text IO (TraceBenchTxSubmit (GenTxId blk))
-  , Transformable Text IO (SendRecvTxSubmission blk)
-  )
-
-createTracers
-  :: forall blk
-  .  BenchTraceConstraints blk
-  => LoggingLayer
-  -> BenchTracers IO blk
-createTracers loggingLayer =
-  BenchTracers
-    baseTrace
-    benchTracer
-    connectTracer
-    submitTracer
-    lowLevelSubmitTracer
-    n2nSubmitTracer
- where
-  baseTrace :: Trace IO Text
-  baseTrace = llBasicTrace loggingLayer
-
-  tr :: Trace IO Text
-  tr = llAppendName loggingLayer "cli" baseTrace
-
-  tr' :: Trace IO Text
-  tr' = appendName "generate-txs" tr
-
-  benchTracer :: Tracer IO (TraceBenchTxSubmit (GenTxId blk))
-  benchTracer = toLogObjectVerbose (appendName "benchmark" tr')
-
-  connectTracer :: Tracer IO SendRecvConnect
-  connectTracer = toLogObjectVerbose (appendName "connect" tr')
-
-  submitTracer :: Tracer IO (SendRecvTxSubmission blk)
-  submitTracer = toLogObjectVerbose (appendName "submit" tr')
-
-  lowLevelSubmitTracer :: Tracer IO TraceLowLevelSubmit
-  lowLevelSubmitTracer = toLogObjectVerbose (appendName "llSubmit" tr')
-
-  n2nSubmitTracer :: Tracer IO NodeToNodeSubmissionTrace
-  n2nSubmitTracer = toLogObjectVerbose (appendName "submit2" tr')
-
--- TODO: move out to Types.hs
-paramsConfig :: Params blk -> TopLevelConfig blk
-paramsConfig (ParamsByron   c _) = c
-paramsConfig (ParamsShelley c _) = c
-
--- * Submission
---
-submitTx' :: (RunNode blk)
-         => Params blk
-         -> IOManager
-         -> SocketPath
-         -> GenTx blk
-         -> IO ()
-submitTx' p iocp sockpath tx =
-    let path = unSocketPath sockpath
-        muxTracer :: Show peer => Tracer IO (WithMuxBearer peer MuxTrace)
-        muxTracer = toLogObject $ appendName "Mux" (trBase p)
-        handshakeTracer :: Tracer IO
-                           (WithMuxBearer (ConnectionId LocalAddress)
-                            (TraceSendRecv (Handshake NodeToClient.NodeToClientVersion CBOR.Term)))
-        handshakeTracer = toLogObject $ appendName "Handshake" (trBase p)
-    in NodeToClient.connectTo
-         (NodeToClient.localSnocket iocp path)
-         NetworkConnectTracers {
-             nctMuxTracer       = muxTracer,
-             nctHandshakeTracer = handshakeTracer
-           }
-         (localInitiatorNetworkApplication (trLowLevel p) (paramsConfig p) tx)
-         path
-
-localInitiatorNetworkApplication
-  :: forall blk m .
-     ( RunNode blk
-     , MonadST m
-     , MonadThrow m
-     , MonadTimer m
-     )
-  => Tracer m TraceLowLevelSubmit
-  -> TopLevelConfig blk
-  -> GenTx blk
-  -> Versions NodeToClient.NodeToClientVersion NodeToClient.DictVersion
-              (OuroborosApplication InitiatorMode NodeToClient.LocalAddress ByteString m () Void)
-
-localInitiatorNetworkApplication tracer cfg tx =
-  foldMapVersions
-    (\v ->
-      versionedNodeToClientProtocols
-        (nodeToClientProtocolVersion proxy v)
-        versionData
-        (protocols v))
-    (supportedNodeToClientVersions proxy)
- where
-    proxy :: Proxy blk
-    proxy = Proxy
-
-    versionData = NodeToClientVersionData $ getNetworkMagic $ configBlock cfg
-
-    protocols :: BlockNodeToClientVersion blk
-              -> NodeToClient.ConnectionId NodeToClient.LocalAddress
-              -> Control.Monad.Class.MonadSTM.STM m RunOrStop
-              -> NodeToClient.NodeToClientProtocols InitiatorMode ByteString m () Void
-    protocols byronClientVersion _ _=
-        NodeToClient.NodeToClientProtocols {
-          NodeToClient.localChainSyncProtocol =
-            InitiatorProtocolOnly $
-              MuxPeer
-                nullTracer
-                cChainSyncCodec
-                NodeToClient.chainSyncPeerNull
-
-        , NodeToClient.localTxSubmissionProtocol =
-            InitiatorProtocolOnly $
-              MuxPeerRaw $ \channel -> do
-                traceWith tracer TraceLowLevelSubmitting
-                (result, maybs) <- runPeer
-                                     nullTracer -- (contramap show tracer)
-                                     cTxSubmissionCodec
-                                     channel
-                                     (LocalTxSub.localTxSubmissionClientPeer
-                                       (txSubmissionClientSingle tx))
-                case result of
-                  SubmitSuccess -> traceWith tracer TraceLowLevelAccepted
-                  SubmitFail msg -> traceWith tracer (TraceLowLevelRejected $ show msg)
-                return ((), maybs)
-
-        , NodeToClient.localStateQueryProtocol =
-            InitiatorProtocolOnly $
-              MuxPeer
-                nullTracer
-                cStateQueryCodec
-                NodeToClient.localStateQueryPeerNull
-        }
-     where
-      Codecs { cChainSyncCodec
-             , cTxSubmissionCodec
-             , cStateQueryCodec
-             } = defaultCodecs (getCodecConfig $ configBlock cfg) byronClientVersion
-
-
--- | A 'LocalTxSubmissionClient' that submits exactly one transaction, and then
--- disconnects, returning the confirmation or rejection.
---
-txSubmissionClientSingle
-  :: forall tx reject m.
-     Applicative m
-  => tx
-  -> LocalTxSub.LocalTxSubmissionClient tx reject m (LocalTxSub.SubmitResult reject)
-txSubmissionClientSingle tx = LocalTxSub.LocalTxSubmissionClient $
-    pure $ LocalTxSub.SendMsgSubmitTx tx $ \mreject ->
-      pure (LocalTxSub.SendMsgDone mreject)
-
-data NodeToNodeSubmissionTrace
-  = ReqIdsBlocking  Ack Req
-  | IdsListBlocking Int
-
-  | ReqIdsPrompt    Ack Req
-  | IdsListPrompt   Int
-
-  | ReqTxs          Int
-  | TxList          Int
-
-  | EndOfProtocol
-  | KThxBye
-
-instance ToObject NodeToNodeSubmissionTrace where
-  toObject MinimalVerbosity = const emptyObject -- do not log
-  toObject _ = \case
-    ReqIdsBlocking  (Ack ack) (Req req) ->
-                               mkObject [ "kind" .= A.String "ReqIdsBlocking"
-                                        , "ack"  .= A.toJSON ack
-                                        , "req"  .= A.toJSON req ]
-    IdsListBlocking sent    -> mkObject [ "kind" .= A.String "IdsListBlocking"
-                                        , "sent" .= A.toJSON sent ]
-    ReqIdsPrompt    (Ack ack) (Req req) ->
-                               mkObject [ "kind" .= A.String "ReqIdsPrompt"
-                                        , "ack"  .= A.toJSON ack
-                                        , "req"  .= A.toJSON req ]
-    IdsListPrompt   sent    -> mkObject [ "kind" .= A.String "IdsListPrompt"
-                                        , "sent" .= A.toJSON sent ]
-    EndOfProtocol           -> mkObject [ "kind" .= A.String "EndOfProtocol" ]
-    KThxBye                 -> mkObject [ "kind" .= A.String "KThxBye" ]
-    ReqTxs          req     -> mkObject [ "kind" .= A.String "ReqTxs"
-                                        , "req"  .= A.toJSON req ]
-    TxList          sent    -> mkObject [ "kind" .= A.String "TxList"
-                                        , "sent" .= A.toJSON sent ]
 
 {-------------------------------------------------------------------------------
   Parametrisation & state
@@ -647,18 +141,6 @@ mkSubmission sTrace sParams@SubmissionParams{spTargets=sThreads, spQueueLen} =
 {-------------------------------------------------------------------------------
   Results
 -------------------------------------------------------------------------------}
-
--- | Summary of a submission run.
-data SubmissionSummary
-  = SubmissionSummary
-    { ssTxSent        :: !Sent
-    , ssTxUnavailable :: !Unav
-    , ssElapsed       :: !NominalDiffTime
-    , ssEffectiveTps  :: !TPSRate
-    , ssThreadwiseTps :: ![TPSRate]
-    } deriving (Show, Generic)
-instance ToJSON SubmissionSummary
-
 data SubmissionThreadStats
   = SubmissionThreadStats
     { stsAcked         :: {-# UNPACK #-} !Ack
@@ -767,14 +249,14 @@ txSubmissionClient
      , txid ~ GenTxId block
      , tx ~ GenTx block
      )
-  => Tracer m (TraceBenchTxSubmit txid)
-  -> Tracer m NodeToNodeSubmissionTrace
+  => Tracer m NodeToNodeSubmissionTrace
+  -> Tracer m (TraceBenchTxSubmit txid)
   -> Submission m block
   -> Natural
   -- This return type is forced by Ouroboros.Network.NodeToNode.connectTo
   -> TxSubmissionClient txid tx m ()
 txSubmissionClient
-    bmtr tr
+    tr bmtr
     sub@Submission{sReportsRefs}
     threadIx =
   TxSubmissionClient $ do
